pyproject.toml
[project]
name = "kindsealion"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Taylor Monacelli", email = "taylormonacelli@gmail.com" }
]
dependencies = [
    "jinja2>=3.1.3",
    "ruamel.yaml>=0.18.6",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
kindsealion = "kindsealion:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.2.0",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/kindsealion"]

[tool.ruff.lint]
extend-select = ["I"]
